move mock api client to single test file
