+++ pkg/twitchtsbot/twitch/global_test.go
@@ -1,0 +1,11 @@
+package twitch
+
+import "sync"
+
+type mockApiClient struct {
+	*sync.Mutex
+	returnErr    bool
+	requestCount int
+	statusCode   int
+	streams      map[string]bool
+}
